Add openapi spec for auth
